fix(cache-manager)!: Fix configured cache manager error
Original behaviour: A default cache manager with cache key 'DynamicCachedFontsFontCacheKey' was used which stored font files in a folder with the same name.
If a value for `cacheStalePeriod` or `maxCacheObjects` is provided which differs from the default value then a new instance of `CacheManager` would be created with the generated `cacheKey`.

Problem: Once a garbage collecttion event is run, `_cacheManagers`  is destroyed and hence the package forgets about the modified `CacheManager` and attempts to locate the file in the default cache folder (as mentioned above).

Solution: There will be no default instance of `CacheManager`, instead an instance will be created using `cacheKey` after any garbage collection events (which includes app restarts). The first time a font is stored will be done using the cache config provided and for subsequent retrievals, `CacheManger` with `cacheKey` will be used and so the package will look for the font file in the right place.

See #233

BREAKING CHANGES:

This commit does not include any changes to the public API. However, as mentioned above,
font files which have already been cached under 'DynamicCachedFontsFontCacheKey' will no
longer be used. A migration tool will be provided to delete old font files.